requirements.txt
strands-agents>=1.30.0
strands-agents-tools>=0.2.20
bedrock-agentcore>=1.4.0
boto3>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
aws-opentelemetry-distro>=0.10.0
